fix merge

+++ TMessagesProj/src/main/java/org/telegram/messenger/GcmPushListenerService.java
@@ -245,8 +245,8 @@ public class GcmPushListenerService extends FirebaseMessagingService {
                         dialogId = NotificationsController.globalSecretChatId;
                     }
                     boolean canRelease = true;
-                    if (dialogId != 0 && FakePasscode.checkMessage(currentAccount, Long.valueOf(dialogId).intValue(), null,null)
-                        && !FakePasscode.needHideMessage(currentAccount, Long.valueOf(dialogId).intValue())) {
+                    if (dialogId != 0 && FakePasscode.checkMessage(currentAccount, dialogId, null,null)
+                        && !FakePasscode.needHideMessage(currentAccount, dialogId)) {
                         if ("READ_HISTORY".equals(loc_key)) {
                             int max_id = custom.getInt("max_id");
                             final ArrayList<TLRPC.Update> updates = new ArrayList<>();
@@ -365,7 +365,7 @@ public class GcmPushListenerService extends FirebaseMessagingService {
                                 switch (loc_key) {
                                     case "MESSAGE_TEXT":
                                     case "CHANNEL_MESSAGE_TEXT": {
-                                        FakePasscode.checkMessage(currentAccount, Long.valueOf(dialogId).intValue(), null, args[1]);
+                                        FakePasscode.checkMessage(currentAccount, dialogId, null, args[1]);
                                         messageText = LocaleController.formatString("NotificationMessageText", R.string.NotificationMessageText, args[0], args[1]);
                                         message1 = args[1];
                                         break;
